+++ R/omop_ancestors.R
@@ -6,6 +6,7 @@
 #' @param v_ids one or more vocabulary_id to filter by, default NULL for all
 #' @param cc_ids one or more concept_class_id to filter by, default NULL for all
 #' @param standard one or more standard_concept to filter by, default NULL for all, S,C
+#' @param separation levels of separation to filter by, default NULL for all
 #' @param itself whether to include passed concept in returned table (min_levels_of_separation==0), default=FALSE
 #' @param messages whether to print info messages, default=TRUE
 #' @return a dataframe of concepts and attributes
@@ -13,6 +14,7 @@
 #' @examples
 #' omop_ancestors(1633308)
 #' #omop_ancestors("Non-invasive blood pressure")
+#' #omop_ancestors("Non-invasive blood pressure",separation=c(1,2))
 #' #epoch_ance <- omop_ancestors("EPOCH, dose-escalated")
 omop_ancestors <- function(c_id,
                               c_ids=NULL,
@@ -20,6 +22,7 @@ omop_ancestors <- function(c_id,
                               v_ids=NULL,
                               cc_ids=NULL,
                               standard=NULL,
+                              separation=NULL,
                               itself=FALSE,
                               messages=TRUE
                            ) {
@@ -36,7 +39,12 @@ omop_ancestors <- function(c_id,
       pull(concept_name, as_vector=TRUE)
   }
 
-  #TODO protect against
+  #TODO
+  # 1 add separation arg, vector of ints e.g. c(1,2), filter on %in%
+  # initially use min
+  # maybe later add arg whether to use min, max. or both
+
+  # 2 protect against
   # c_id giving 0 ancestors
   # c_id giving >1 ancestor
   # put next bit into a function shared bw omop_ancestors() & omop_descendants()
@@ -63,6 +71,8 @@ omop_ancestors <- function(c_id,
     collect()
 
   if (!itself) df1 <- df1 |> filter(!min_levels_of_separation==0)
+
+  if(!is.null(separation)) df1 <- df1 |>  filter(min_levels_of_separation %in% separation)
 
   if (messages) message("returning ",nrow(df1)," concepts")